refactored hamburger menu, added in all spec'd icons
